Cleanup OpenLayers geocoder.
Remove unneeded call to address.toProprietary(this.api).
Merge both requests, use a single variable for the Nominatim url address, inline the callback.
Move status check out of geocode_callback.
Fix error messages (because the server is running Nominatim, OpenLayers is used just to make requests.
Do not prepare the address string when not needed (when reverse geocoding is used).