leaflet.core: Add support for overview, pan, and large zoom controls using Control.MiniMap, Control.Zoomslider, and Control.Pan Leaflet plugins.
To use these controls, one has to load the corresponding plugins first.
These plugins are not official, but are listed on http://leafletjs.com/plugins.html "Notable Leaflet Plugins" page.
If the plugins are not loaded, overview and pan are ignored, and large zoom falls back to small zoom, so nothing changed in this case.